+++ manvi/gussetcheck/check_nocgo.go
@@ -1,0 +1,12 @@
+//go:build !cgo
+
+package gussetcheck
+
+import "context"
+
+func check(ctx context.Context) error {
+	if err := ctx.Err(); err != nil {
+		return err
+	}
+	return ErrNotLinked
+}